feat: Store channel opening params on coordinator
